feat: subscription service (#5)
* feat: subscription service setup & routes & handlers

* feat: subscription service - service methods impl & repository interface

* feat: subscription service - firestore repository

* feat: subscription service - firestore repository

* feat: subscription service - aws ses in progress

* feat: subscription service - aws ses done

* feat: subscription service - newsletter service client

* fix: remove unnecessary directory

* fix: logging msg

* ref: marshall, logger, traefic conf, env

* feat: notify subscribers after publishing post in newsletter service

* ref: cleaner newsletter & post repository

* ref: x_user_id as util function

* ref: get newsletter id as util function

* ref: persistence model for subscription

* ref: email provider interface

+++ services/newsletter-service/internal/service/services/impl/post_service.go
@@ -11,14 +11,20 @@ import (
 )
 
 type postService struct {
-	postRepo       repositories.IPostRepository
-	newsletterRepo repositories.INewsletterRepository
+	postRepo                  repositories.IPostRepository
+	newsletterRepo            repositories.INewsletterRepository
+	subscriptionServiceClient services.ISubscriptionServiceClient
 }
 
-func NewPostService(postRepo repositories.IPostRepository, newsletterRepo repositories.INewsletterRepository) services.PostService {
+func NewPostService(
+	postRepo repositories.IPostRepository,
+	newsletterRepo repositories.INewsletterRepository,
+	subscriptionServiceClient services.ISubscriptionServiceClient,
+) services.PostService {
 	return &postService{
-		postRepo:       postRepo,
-		newsletterRepo: newsletterRepo,
+		postRepo:                  postRepo,
+		newsletterRepo:            newsletterRepo,
+		subscriptionServiceClient: subscriptionServiceClient,
 	}
 }
 
@@ -35,10 +41,15 @@ func (p postService) CreatePost(ctx context.Context, post *model.Post, newslette
 		return nil, errors2.ErrInvalidUUID
 	}
 
-	_, err = p.newsletterRepo.GetByIdAndUserId(ctx, parsedNewsletterID, parsedUserID)
+	newsletter, err := p.newsletterRepo.GetById(ctx, parsedNewsletterID)
 	if err != nil {
 		logger.Error("Failed to get newsletter by ID", "newsletterID", newsletterID, "error", err)
 		return nil, errors2.ErrNotFound
+	}
+
+	if newsletter.UserID != parsedUserID {
+		logger.Error("Unauthorized access to create post", "newsletterID", newsletterID, "userID", userID)
+		return nil, errors2.ErrUserNotAuthor
 	}
 
 	post.NewsletterID = parsedNewsletterID
@@ -111,10 +122,21 @@ func (p postService) UpdatePost(ctx context.Context, postID string, newsletterID
 		return nil, errors2.ErrInvalidUUID
 	}
 
-	post, err := p.postRepo.GetByIdAndUserId(ctx, parsedPostID, parsedNewsletterID, parsedUserID)
+	post, err := p.postRepo.GetById(ctx, parsedPostID, parsedNewsletterID)
 	if err != nil {
 		logger.Error("Failed to get post by ID", "postID", postID, "error", err)
 		return nil, errors2.ErrPostNotFound
+	}
+
+	newsletter, err := p.newsletterRepo.GetById(ctx, parsedNewsletterID)
+	if err != nil {
+		logger.Error("Failed to get newsletter by ID", "newsletterID", newsletterID, "error", err)
+		return nil, errors2.ErrNotFound
+	}
+
+	if newsletter.UserID != parsedUserID {
+		logger.Error("Unauthorized access to update post", "postID", postID, "userID", userID)
+		return nil, errors2.ErrUserNotAuthor
 	}
 
 	if postUpdate.Title != nil {
@@ -154,9 +176,20 @@ func (p postService) DeletePost(ctx context.Context, postID string, newsletterID
 		return errors2.ErrInvalidUUID
 	}
 
-	_, err = p.postRepo.GetByIdAndUserId(ctx, parsedPostID, parsedNewsletterID, parsedUserID)
+	_, err = p.postRepo.GetById(ctx, parsedPostID, parsedNewsletterID)
 	if err != nil {
 		return errors2.ErrPostNotFound
+	}
+
+	newsletter, err := p.newsletterRepo.GetById(ctx, parsedNewsletterID)
+	if err != nil {
+		logger.Error("Failed to get newsletter by ID", "newsletterID", newsletterID, "error", err)
+		return errors2.ErrNotFound
+	}
+
+	if newsletter.UserID != parsedUserID {
+		logger.Error("Unauthorized access to delete post", "postID", postID, "userID", userID)
+		return errors2.ErrUserNotAuthor
 	}
 
 	err = p.postRepo.Delete(ctx, parsedPostID, parsedNewsletterID)
@@ -185,10 +218,21 @@ func (p postService) PublishPost(ctx context.Context, postID string, newsletterI
 		return errors2.ErrInvalidUUID
 	}
 
-	post, err := p.postRepo.GetByIdAndUserId(ctx, parsedPostID, parsedNewsletterID, parsedUserID)
+	post, err := p.postRepo.GetById(ctx, parsedPostID, parsedNewsletterID)
 	if err != nil {
 		logger.Error("Failed to get post by ID", "postID", postID, "error", err)
 		return errors2.ErrPostNotFound
+	}
+
+	newsletter, err := p.newsletterRepo.GetById(ctx, parsedNewsletterID)
+	if err != nil {
+		logger.Error("Failed to get newsletter by ID", "newsletterID", newsletterID, "error", err)
+		return errors2.ErrNotFound
+	}
+
+	if newsletter.UserID != parsedUserID {
+		logger.Error("Unauthorized access to publish post", "postID", postID, "userID", userID)
+		return errors2.ErrUserNotAuthor
 	}
 
 	if post.Published {
@@ -196,7 +240,17 @@ func (p postService) PublishPost(ctx context.Context, postID string, newsletterI
 		return errors2.ErrAlreadyPublished
 	}
 
-	//TODO: call subscription service to notify subscribers
+	notification := &services.Notification{
+		Title:       post.Title,
+		Content:     post.Content,
+		HtmlContent: post.HtmlContent,
+	}
+
+	err = p.subscriptionServiceClient.NotifySubscribers(ctx, newsletterID, notification)
+	if err != nil {
+		logger.Error("Failed to notify subscribers", "newsletterID", newsletterID, "error", err)
+		return err
+	}
 
 	post.Published = true
 	_, err = p.postRepo.Update(ctx, post)
